Add files via upload
Living with a hearing impairment negatively affects the person with it as well as those around them. To lessen this detrimental effect,
many apps and technical instruments have been developed. most of the mobile apps that have been created as a result, not all types of
hearing challenged people can utilize speech-to-text technology, it only functions in certain predetermined situations, and it does not
enable conversations involving several participants.

As a result, the current tools are less useful, and people who are hard of hearing feel like they aren't fully engaged in the discourse.

In order to solve this, the model presented in this research introduces the use of Multiple-Speaker Classification technology in the
creation of mobile applications for hearing challenged users.

Additionally, we demonstrate the Deaf Chat prototype, a mobile application that makes use of the newly created model.

The potential of this application to meet the needs of those who are hard of hearing was assessed using a survey.

The evaluation findings showed positive user acceptability and demonstrated that a platform like Deaf Chat might be beneficial for
people with hearing loss as a whole.